refactor: rename perfil component files
